Compile assembly for 'any CPU' instead of 'x86'
For some reason, the app was compiling for 'x86' platforms intead of 'any CPU' by default. The registry import requires the app to be running as an x64 app on an x64 operating system to write to the correct keys. When running as an x86 process on an x64 operating system, the x86 version of reg.exe gets called instead of the x64 version which results in a WOW6432Node key getting placed between the HKLM\Software and SecureW2 keys when performing the reg file import.

The following instructions were followed to adjust the target platform in Visual C# 2010 Express:
* http://forums.xna.com/forums/t/4377.aspx#22601
* http://www.dpvreony.co.uk/blog/post/v/62