Replace non-working calls to Application.Exit() with Environment.Exit(1)
Since the application does not start a message loop, Application.Exit() does not work as expected. Using Environment.Exit(1) will cause the application to exit with an exit code of 1. This application is more of a console app with a GUI overlayed.

With this modification, keep an eye out for possible Security exceptions.